Removes old QPE files.
